cleans compiler warnings with cast

Closes #139

See merge request fun3d-developers/refine!905

+++ src/ref_cavity.c
@@ -943,13 +943,14 @@ REF_FCN REF_STATUS ref_cavity_form_insert(REF_CAVITY ref_cavity,
   }
 
   if (ref_cavity_debug(ref_cavity))
-    printf("insert form state %d\n", ref_cavity_state(ref_cavity));
+    printf("insert form state %d\n", (int)ref_cavity_state(ref_cavity));
   RSS(ref_cavity_verify_face_manifold(ref_cavity), "ball face manifold");
   if (ref_cavity_debug(ref_cavity))
-    printf("insert face manifold state %d\n", ref_cavity_state(ref_cavity));
+    printf("insert face manifold state %d\n",
+           (int)ref_cavity_state(ref_cavity));
   RSS(ref_cavity_verify_seg_manifold(ref_cavity), "ball seg manifold");
   if (ref_cavity_debug(ref_cavity))
-    printf("insert seg manifold state %d\n", ref_cavity_state(ref_cavity));
+    printf("insert seg manifold state %d\n", (int)ref_cavity_state(ref_cavity));
 
   return REF_SUCCESS;
 }
@@ -1025,10 +1026,10 @@ REF_FCN REF_STATUS ref_cavity_form_insert2(REF_CAVITY ref_cavity,
   if (ref_cavity_debug(ref_cavity)) ref_cavity_tec(ref_cavity, "form-tri.tec");
 
   if (ref_cavity_debug(ref_cavity))
-    printf("insert form tri state %d\n", ref_cavity_state(ref_cavity));
+    printf("insert form tri state %d\n", (int)ref_cavity_state(ref_cavity));
   RSS(ref_cavity_verify_seg_manifold(ref_cavity), "ball seg manifold");
   if (ref_cavity_debug(ref_cavity))
-    printf("insert tri manifold state %d\n", ref_cavity_state(ref_cavity));
+    printf("insert tri manifold state %d\n", (int)ref_cavity_state(ref_cavity));
 
   RSS(ref_cavity_enlarge_conforming(ref_cavity), "enlarge boundary");
   if (REF_CAVITY_VISIBLE != ref_cavity_state(ref_cavity)) return REF_FAILURE;
@@ -1084,10 +1085,11 @@ REF_FCN REF_STATUS ref_cavity_form_insert2(REF_CAVITY ref_cavity,
   if (ref_cavity_debug(ref_cavity)) ref_cavity_tec(ref_cavity, "form-tet.tec");
 
   if (ref_cavity_debug(ref_cavity))
-    printf("insert face form state %d\n", ref_cavity_state(ref_cavity));
+    printf("insert face form state %d\n", (int)ref_cavity_state(ref_cavity));
   RSS(ref_cavity_verify_face_manifold(ref_cavity), "ball face manifold");
   if (ref_cavity_debug(ref_cavity))
-    printf("insert face manifold state %d\n", ref_cavity_state(ref_cavity));
+    printf("insert face manifold state %d\n",
+           (int)ref_cavity_state(ref_cavity));
 
   return REF_SUCCESS;
 }
